requirement.txt
pandas
tensorflow
numpy
opencv-python
gym==0.25.2
matplotlib
tqdm
torch
seaborn
pyyaml
ffmpeg-python